clean code

+++ bienvakieudulieu/bai9.cpp
@@ -1,11 +1,11 @@
-#include <iostream>
+#include <bits/stdc++.h>
+
 using namespace std;
 
-int main (){
-    int a = 12984019;
-    int b = 12948120;
-    cout << (long long)a * b;
-    
+int main()
+{
+    long long a = 12984019;
+    long long b = 12948120;
+    cout << a * b;
     return 0;
-
 }
